Updated course management apis with program id

+++ coursemanagement/src/main/java/com/example/coursemanagement/controller/ProgramController.java
@@ -1,11 +1,15 @@
 package com.example.coursemanagement.controller;
 
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.example.coursemanagement.model.Program;
 import com.example.coursemanagement.service.ProgramService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/program")
@@ -20,7 +24,7 @@ public class ProgramController {
 
     // Get a program by programId
     @GetMapping("/{programId}")
-    public Program getProgramById(@PathVariable Long programId) {
+    public Program getProgramById(@PathVariable Integer programId) {
         // Fetch the program details by programId using the ProgramService
         return programService.getProgramById(programId);
     }
